% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bday_calculation}
\alias{bday_calculation}
\title{bday_calculation}
\usage{
bday_calculation(station, year = "all")
}
\arguments{
\item{station}{a character representing the station number that exists in hydat}

\item{year}{a vectoer of numeric representing the list of years used for calculating
the threshold}
}
\value{
a dataframe of the format || station || year || lastbday || firstbday ||
}
\description{
bday_calculation calculates the last and first bday for some years with
record at a station.
}
