% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Get_ann_mean_flow}
\alias{Get_ann_mean_flow}
\title{Get_ann_mean_flow}
\usage{
Get_ann_mean_flow(station, year = "all")
}
\arguments{
\item{station}{a character indicating the station number to calculate the annual mean flow;}

\item{year}{a character indicating the list of years where the annual mean flow is being
calculated, or by default equals to 'all' indicating all recorded years at the station}
}
\value{
a data frame of the format:
|| station || Year || annual mean yield||
}
\description{
Get_ann_mean_flow calculates the annual mean flow based on a station and a year
vector indicating which year the mean flow will be calculated.
}
