% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Get_Drought_metrics}
\alias{Get_Drought_metrics}
\title{Get_drought_metrics}
\usage{
Get_Drought_metrics(station, year = "all")
}
\arguments{
\item{station}{a character representing a station we want the flood metrics to be calculated at;}

\item{year}{a vector of numeric representing the list of years we want thew flood metrics to be calculated at,
defaulted to 'all' indicating all the years at the station;}
}
\value{
A dataframe of the format:
|| STATION_NUMBER || Year || dr_days || dr_events || dr_threshold || dut_max_dur ||
}
\description{
The function takes in a character representing the station number within 1014
RHBN_U stations and a vector of numerics or a character, calculates four drought metrics:
dr_threshold: the lower threshold that was calculated in Threshold_build function from above;
dr_days:days under the lower threshold within a year;
dr_events:independent drought events over a year;
dut_ma_dur: the maximum consecutive duration under the threshold within a year;
}
\details{
The four metrics will then be calculated based on only the time within a year between the
last and first bday calculated from the previous bday_calculation() function,
and the result will be recorded in a table and returned.
In order to measure the independence between drought events, we set the following
constraints:
1. t >= 5
2. the excess volume, Vabove / min(Vbelowi, Vbelowj)>=0.1
where t is the time between events, Vabove is the amount that exceeds the threshold,
Vbelowi and Vbelowj are the volumes below during events i and j.
}
\references{
WMO World Meteorological Organization (2008). Manual on Low-Flow Estimation and Prediction. 1339 Operational Hydrology.
}
