% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Get_flood_metrics}
\alias{Get_flood_metrics}
\title{Get_flood_metrics}
\usage{
Get_flood_metrics(station, year = "all")
}
\arguments{
\item{station}{a character representing a station we want the flood metrics to be calculated at;}

\item{year}{a vector of numeric representing the list of years we want thew flood metrics to be calculated at,
defaulted to 'all' indicating all the years at the station;}
}
\value{
A dataframe of the format:
|| STATION_NUMBER || Year || pot_days || pot_events || pot_threshold || pot_max_dur ||
}
\description{
Get_flood_metrics takes in a character representing the station number within 1014
RHBN_U stations and a vector of numerics or a character, calculates four flood metrics:
pot_threshold: the upper threshold that was calculated in Threshold_build function from above;
pot_days:days over the threshold within a year;
pot_events:independent flood events over a year;
pot_ma_dur: the maximum consecutive duration over the threshold within a year;
The four metrics will then be recorded in a table and returned.
}
\details{
Note that there are requirements regarding the independence conditions for pot_event metric,
we implement a check for the correlation between flood events.
Referring to Lang etal (1999), we set two conditions:
1) R > 5+log(A/1.609^2);
2) Xmin < 0.75 * min(Xi, Xj);
where R is the time in days between two peaks, A is the watershed area in squared kilometers;
Xmin is the minimum flow between the adjacent flow peaks Xi and Xj.
In order for two seperate peaks to be considered as independent events of drought or flood,
both conditions need to be met. If either of the conditions fail, we consider two events
as the same one. R is defined within the iteration.
}
\references{
Slater et al.(2016). On the impact of gaps on trend detection un extreme
streamflow time series. Int.J.climatol. 37:3976-3983(2017). doi: 10.1002/joc.4954
}
