% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{percentile.ranked}
\alias{percentile.ranked}
\title{percentile.ranked}
\usage{
percentile.ranked(a.vector, value)
}
\arguments{
\item{a.vector}{a vector of any comparable data type, ideally numeric;}

\item{value}{a comparable data type to the components in a.vector, ideally numeric}
}
\value{
the percentage of values within the vector that's smaller than the value itself.
}
\description{
Calculates the pencentage of values in a vector smaller than value
}
