% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Threshold_Build}
\alias{Threshold_Build}
\title{Threshold_Build}
\usage{
Threshold_Build(
  stations,
  yfrom = 1981,
  yto = 2010,
  hi = 0.95,
  low = 0.05,
  write_csv = FALSE
)
}
\arguments{
\item{stations}{a vector of characters indicating the list of station we want to calculate the
threshold on;}

\item{yfrom}{a numeric indicating the year threshold calculation starts, defaulted to 1981;}

\item{yto}{a numeric indicating the year threshold calculation ends, defaulted to 2010;}

\item{hi}{a numeric indicating the upper confident threshold, defaulted to 0.95;}

\item{low}{a numeric indicating the lower confident threshold, defaulted to 0.05}

\item{write_csv}{a boolean value indicating whether ot not the output will be written
to csv, defaulted to FALSE;}
}
\value{
A dtaaframe of the format:
|| station ||UpperThreshold ||LowerThreshold ||
}
\description{
Threshold_Building takes into a list of stations, year_from and year_to representing the beginning
and the end of the reference period, and the two ratio numbers hi and low representing the
percentile(high and low). bday_calculation function will be needed to calculate the threshold.
}
